feat: 내가 가입한 모임 목록 조회
